backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
tiktoken==0.8.0
httpx==0.28.1
python-dotenv==1.0.1
